+++ src/util/math.hpp
@@ -3,7 +3,7 @@
 namespace myrisa {
 namespace util {
 
-inline float rescale(float x2, float x1, float y2, float y1) {
+inline double rescale(double x2, double x1, double y2, double y1) {
   return (x2 - x1) / (y2 - y1);
 }
